feat: auto-determine nav item active state
It will make the the `db-navigation-item`'s `active` prop  optional.
It highlights an active link element with it's children.
It still works as before but in addition it checks for child elements.
They styles like active elements when having the `aria-current` attr set.
It allows to use framework build-in aria-current mechanisms.

closes #2200